[refactor] 상품 목록 페이지 코드 리팩토링
